Implementation of hot-reload for both frontend and backend services through Docker-Compose
Due to issues with turbopack in container environments that have not been resolved by Vercel, it was necessary to revert to webpack polling to update file changes automatically. The same was true for nodemon automatically reloading the Express app; using polling with legacy-watch flag. Custom poll intervals for both services.